requirements.txt
discord.py==1.2.2
python-dotenv==0.10.3
PyNaCl==1.3.0
git+https://github.com/Thetoto/pytube
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-texttospeech==0.4.0
google-api-python-client==1.7.9